Hide warning when reconfiguring upstream from local to remote branch
When using a local branch as upstream, the branch creation dialog showed
a warning that using a local branch as upstream is not recommended. This
warning was not removed when reconfiguring upstream to a remote branch.

This regression was introduced with the fix for bug 440097. That one
removed the line "setMessage(null)". This fix reintroduces that line,
but as "setMessage(initialMessageWhenThePageWasCreated)" instead. This
avoids the previously experienced wrong height calculation of the
dialog, I have tested that extensively.

Bug: 452212
Change-Id: I849c483b5a8a42a80191c4105b102fe4cd5377c4
Signed-off-by: Michael.Keppler <Michael.Keppler@gmx.de>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>