Provide a way to configure RepositoryChangeScanner interval
RepositoryChangeScanner was triggered every 10 seconds if the UI was
active. This is an overkill and causes unneeded CPU load by idling
Eclipse. This change increases the interval to 5 minutes, adds new
"Refresh interval" preference to the root Git preference page and
changes the job in the way that it listens to the preference change. If
the preference is set to 0, the job stops automatic repositories change
check.

Change-Id: I02751334d4cb6dea97f575d24a5fa24c6832eee4
Signed-off-by: Andrey Loskutov <loskutov@gmx.de>