Fix representation toggle commands in the repositories view
Make the ToggleBranchCommit and the ToggleBranchHierarchy commands
work also when invoked (for instance via Quick Access) when the
currently active view is not the repositories view.

These two commands use default handlers; when invoked via Quick
Access for instance when the Outline view was the active view,
they would produce an ExecutionException since no active view
could be determined. The commands would thus execute halfway
through: they'd update the command's state, but then fail and
instead of refreshing the view would pop up an error dialog.

I see no reason to restrict those two commands to be available
only when the repositories view is the active view. So instead
of eliminating the default handler, change the implementation
such that the commands work in any case.

Bug: 495064
Change-Id: Ib6b0c9fd38940efd6a15c0a0fd20be2a247ee7ec
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>