Throttle updates of the EclipseGitProgressTransformer
This progress monitor updated its SubMonitor every time if
total == UNKNOWN. This can be very inefficient since it may
involve a UI update. Oxygen seems to include a progress monitor
that optimizes for this, but on Neon and earlier Eclipses,
this progress reporting may slow down operations. Most noticeably
it occurred in the "counting objects" phase of a push operation,
which would become unuseably slow if run in the foreground, but
which would terminate quickly when run in the background.

Bug: 516276
Change-Id: Ib3e37918f493da6a04eb51bcc7086ecd3eb2bc6f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>