Staging view: add toolbar buttons to stage/unstage all files
The buttons do not appear in the contextual action tooltip. The
contextual actions operate on a selection, whereas the new
"stage/unstage all" actions also operate on unselected files.
Including it in the tooltip would give a confusing UI.

Also, if this action were in the tooltip, a slight mis-click would
stage/unstage not only the selected files but all of them. Although
that would not be a severe problem (the changed files are selected
after the operation in the target viewer, so undoing a "stage all"
is as easy as using "unstage seleted files"), let's try to avoid this.

Bug: 518213
Change-Id: I127dae880c28f851b078736a7940dc698ab42faa
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>