FetchGerritChangePage: supply patch set number in the background
Allow to finish the dialog even if only a change number is given.
If so, determine the highest patch set number for that change at
the beginning of the fetch job. If no such change exists, display an
error message, otherwise fetch the latest patch set.

The branch name suggestion for a change ref without patch set number
is constructed using the current time in place of the missing patch
set number. This gives a reasonable default name that also has a high
probability of not existing yet.

Bug: 518492
Change-Id: I580123a1713a64ba6d2cf45dc0b488a4bc97c044
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>