Git history: preference-based actions must run in the UI thread.
Change BooleanPrefAction in GitHistoryPage to run actions in response
to preference changes in the UI thread.

Adapt FetchGerritChangePage; the logic was wrong there regarding the
setting of a preference. Preferences can be changed from any thread
(EclipsePreference is thread-safe). Code modifying a preference should
_not_ need to be aware of any listeners anywhere that might update UI:
reacting to preference changes in the correct thread is clearly the
responsibility of the listeners. The comment in the javadoc of
IPreferenceStore.addPropertyChangeListener() is simply bogus. 

Change-Id: I8e13ee8a700e4cc98129ba00b0982cd3afeb6634
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>