Added context actions 'Assume Unchanged' and 'Untrack'
Modified unstaged changes can be marked as 'Assume Unchanged' or
untracked using a context menu action.

AssumeUnchangedOperation and UntrackOperation are extended by a
constructor that takes a path list. Reason is that it cannot be assumed
that in the StagingView the StagingEntry's are workspace resources. In
this case the selection is a collection of paths of the current selected
repository.

Compact Tree representation and filtered views are also considered.

Bug: 381158
Change-Id: I4eec5d307ae7b1a04ad288b3c3e769218a98a0b5
Signed-off-by: Karsten Thoms <karsten.thoms@itemis.de>