Minor fix in Gerrit version check
According to the documentation,[1] one form of the output of Gerrit's
"gerrit-version" ssh command is "<tagname>-<n>-g<sha1>", where <sha1>
is defined by the documentation as "the seven-character abbreviated
SHA-1 of the commit."

Well, current Gerrit Gerrit reports at the bottom of that very same
documentation page at [1] its version as "2.15-rc1-577-g6b1fed0abe",
which is a 10-character abbreviated SHA1.

So relax the version check to match 7 or more characters.

[1] https://gerrit-review.googlesource.com/Documentation/cmd-version.html

(Such a long commit comment for adding a single comma!)

Change-Id: Id082cc9189a347fe58a2be741256c678bd091f12
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>