Better align repository hyperlink in commit viewer
This removes the parent composite of the repository hyperlink label of
the commit viewer toolbar. The improved layout is just because of not
having the default spacing/margins/borders of the removed parent.

Therefore an alternative fix would be to just set those layout
properties instead of removing the composite, should some developer see
this composite as necessary due to another reason unknown to me.

Bug: 527454
Change-Id: I0694a04c43d92822e762f28d8b3b154ac53441f8
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>