Upgrade commons-compress to 1.15 (used by JGit)
Also update xz to 1.6 as this is the version commons-compress 1.15
wants. Since xz 1.6 is an optional dependency for commons-compress we
need to add a non-optional dependency to xz for
org.eclipse.jgit.pgm.test since one of the tests explicitly requires xz.

Related change adding commons-compress to Orbit:
https://git.eclipse.org/r/#/c/115366/

CQ: 15356
CQ: 15360
Change-Id: If49b5de905a18d16cd66288aaafdc07e92196bb4
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>