Repositories view blocks Eclipse on startup
When asking the "Tags" node whether it has children or not, answer true,
as long as the ref database has not been read, and start reading the
database in a job.

This avoids blocking the Eclipse UI for reading the ref database very
early when creating the nodes in the repository tree view. Of course the
blocking still happens if the user tries to expand the tags node
immediately after that node became visible.

Change-Id: I553db27d2bdd2571c6a713dabd6e68065873b98b
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>
Bug: 533230