Reload commit info in history only if selection really changed
On a SelectionChangedEvent from the commit graph table we re-load
the commit viewer and the file diffs. However, such an event is
also received on Windows when the focus changes back to the commit
graph table and the same commit is selected. Happens for instance also
on a right click on the commit to bring up the context menu.

When we're already showing information for that commit, there is no
need to re-load it.

Problem observed on Windows 10, all Eclipse versions at least since
Neon 3 -- didn't try earlier versions.

Change-Id: I0b9c9b1aa0906b744b8a0144b5d3d08d2fb73100
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>