Fix error marker updates in staging view
Use (absolute) paths instead of resources to determine which elements
to update in the staging viewer. Using resources may give inconsistent
results if

* a file occurs several times in the Eclipse resource tree, or
* a on-disk directory doesn't exist as a resource in Eclipse at all.

The latter can occur if tree representation is used and an Eclipse
project lives in some sub-directory (for instance, "bundles") in the
git working tree. In that case the staging view will have a folder
"bundles", but there's no resource for it, and thus our top-down
tree traversal will skip it and all its descendants.

Bug: 536471
Change-Id: I46fc84737471e2994af056ecb238b3bbfb55e413
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>