History table: let the user define which columns are shown
Columns in tables cannot be shown/hidden individually. But we can
set the width to zero to hide a column. Of course the user can hide
and restore columns directly by dragging the column separators, too.

Add a "Columns" sub-menu in the view menu where the user can check
the columns to be shown/hidden. These modifications are transient;
they are not saved or restored when the view is closed or opened,
and they do not survive across and IDE restart. This is similar to
the "pinned" state of a history view.

To permanently change which columns are shown by default, there's
a new preference page.

Bug: 339796
Change-Id: Ib8daff29b6610eed1397969a203947e5e72b629a
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>