Fetch from Gerrit: if no repository, let the user select one
Replaces the message that no repository is selected if the "Fetch
Change from Gerrit" wizard is invoked without a repository selected
by a selection dialog in which the user can select which Gerrit-
enabled repository he'd like to fetch from.

If there are no Gerrit repositories to choose from show an error
message.

Bug: 507733
Change-Id: I1b8adb4ecf1b769d249f1d7180d6f4a0473131f4
Signed-off-by: Jonas Hungershausen <jonas.hungershausen@vogella.com>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>