Fetch from Gerrit: handle double-click in repo selection dialog
The GitSelectRepositoryPage already would move on to the next page
in the wizard on a double-click on a repository. Expand that to
finish the dialog if there is no next page (and the wizard _can_
be finished).

Also re-order the control creation such that the setup of the viewer
is all in one place. (Except setting its input, which happens later).

Change-Id: I1ac5eaaf4a093d62eb2d55e00931c003c177cdd9
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>