Reflog view: simplify initial selection setting
Remove the new code added in commits c7940ec and 5ae9047 again. We
already call UIUtils to set the initial selection, so there should
be no need to do this again. Fix the method in UIUtils instead to use
the active editor if there is no otherwise active part.

Turns out that the active editor may be set even if there is no active
part, and may not even be active... -- but if it exists at all, it's
the editor the user sees, which is a very reasonable selection source.

Change-Id: I0beb5e87932acd0a8fe48e2702c07b8cb2adb8f2
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>