Staging view: simplify handling of initial selection
There's no need to remember the selection when the view is initialized;
it is sufficient to take the current selection when the view's controls
are created.

When started with "link with selection" off, the view may still come
up without repository. That's consistent with other views such as the
history view. When the user then switches on "link with selection",
populate the view with either the last tracked selection or if there
is none with the selection from the currently active editor, if any.

Bug: 539799
Change-Id: Ie2440f09ba17611b0ace3c0c69934da824e53ace
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>