Change LoginDialog default button label
The login dialog's default button always had the label "Save". This
feels unnatural when the dialog is used for logging in; it may be
fine when the dialog is used to change an existing password.

Change the label to "Log in" in changeCredentials == false. To ensure
this flag is always set properly when we create the buttons, remove
the setter for that field and pass the value as a parameter in the
constructor.

Bug: 541312
Change-Id: I6df0b8f592ba3f91dbb89a83e8a0be30d5738501
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>