Avoid triggering unintended commands
Commit 56991ad introduced the ability to switch repositories in all
EGit views. This works through providing a "fake" selection containing
the repository, and all our views reacting when the current selection
changes and has a different repository. To avoid creating multiple
entries in the Git history view's view history, the code used a
RepositoryNode to encapsulate the repository in that selection.

However, that had the side-effect of enabling commands from the
repository view that should be active only if the real selection
is a RepositoryNode.

Avoid this by using a different class. It's still a RepositoryTreeNode
and thus still compares equal() to a RepositoryNode (which is the
criterion to avoid duplicate view history entries in the history view),
but since it is not a RepositoryNode, commands are not enabled.

For the few cases where command handlers are active on general
RepositoryTreeNodes add conditions in plugin.xml to ensure it's not
such a new RepositoryVirtualNode.

Bug: 541752
Change-Id: Id33727a1adfba44c3bfe781297cbf04032f60458
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>