Style read-only text field like a label
In the SimpleConfigure{Fetch|Push}Dialog, style the URI read-only
text field like a label: remove the border, and change the background
to the one of its parent. Use StyledText instead of Text to be able
to suppress the text caret on GTK. Fix the vertical text alignment,
too.

The result is very much like a selectable label. The text can be
scrolled horizontally, and can be selected and copied to the clipboard.

Bug: 541813
Change-Id: I3c6717f77d713a51def0244602b79ff2933d88b6
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>