[repo view] Introduce repository groups
In order to allow the user to structure repositories a new node type
'group' is introduced. A repository may belong to at most one group.
Groups can be created, renamed and deleted; they cannot be nested. If a
repository belongs to a group, it is not shown in the tree root.
Assignment is done by having repositories selected while creating the
group, by drag/drop or via a new context menu entry.

CQ: 21019
Bug: 545123
Change-Id: I7bf62760f0f242f2f72db1e674ed26141964fde4
Signed-off-by: Alexander Nittka <alex@nittka.de>