Get rid of the RebasePlanIndexer
When a rebase plan is re-parsed due to a repository event (ref
changed, or index diff changed) it was possible that the plan
indexer was updated while the TreeViewer still did a refresh of
items from the previous plan. This would require a thread-safe
indexer, and moreover we would need to make sure that the current
plan and the current indexer always were in synch. From code
inspection I suspect that was not true in all cases.

This plan indexer existed only to provide the step numbers in
the "Step" column in the view. We can add this information to
the PlanElements in the content provider instead, and then
remove the whole indexer. That avoids this whole problem since
then the label provider depends only on the PlanElements actually
in the tree.

Bug: 547434
Change-Id: Ifeefa55090285c464fe912e924d91757dbac6e37
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>