History: show whitespace in the diff viewer
Give the diff viewer in the history view a little context menu, shown
only when there is actually some content. The context menu contains
copy, select all, and show whitespace. The latter enables the user to
switch on and off showing whitespace, and it makes the viewer respect
the global editor preference for showing whitespace.

Bug: 462509
Change-Id: Iad75ffb10d539f9913a85a6d1382bb6ef307511b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>