Repositories view: check for submodules when reacting on selection
When the repositories was set to follow the selection and a file in
a submodule folder was selected, the submodule repository was added
as a top-level repository in the view. Fix this by also considering
repositories under submodule nodes when determining which node to
show.

Add a new test, and fix SubmoduleFolderTest to properly commit the
addition of the submodule in setUp().

Bug: 520735
Change-Id: I54c1d2c5d2b8fac95dc7a4fed658e59285639f4c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>