Always check for merge commit when opening file diff
The 2-way file diff is invoked from 3 different places in the code, but
only one of them checked whether the diff was a merge commit. Move that
check into the file diff method itself to have all 3 invocations
checked.

It might be possible that the other 2 invocations cannot be called for
merge commits due to other external logic, but let's be fault tolerant
here by always forbidding this.

Change-Id: Ieaacc0c0ddf0bdbc273891cd557ab55ff5560133
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>