Remove duplicate configure fetch/push commands
The commands "configure fetch" and "configure upstream fetch" use the
same handler and the same label, therefore one of them is superfluous.
Remove the one without ".team" in the command id, since ".team" is a
convention in the list of command ids in the egit API. Redirect menus
and command images to the remaining command id.

Bug: 549997
Change-Id: I5f8129670566e2f41051f0f533dcf36947d4dd68