Ensure staging view widgets get re-enabled
Make sure the staging view widgets, which get disabled before a
commit, get re-enabled if something goes wrong and a RuntimeException
is thrown. See bug 550336.

As far as I see the only possibility where the widgets could remain
disabled is if the scheduled job never terminates.

If a RuntimeException does occur log and show the error to the user.
Previously it was apparently only logged.

Bug: 550513
Change-Id: I74b2268666a2e441a76a48d462e25f119380760b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>